Updated config.
