+++ src/main/java/com/revature/rideforce/user/UserApplication.java
@@ -1,11 +1,6 @@
 package com.revature.rideforce.user;
 
-import java.io.IOException;
-import java.text.ParseException;
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.InitializingBean;
@@ -20,23 +15,11 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Scope;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.cognitoidp.AWSCognitoIdentityProvider;
-import com.amazonaws.services.cognitoidp.AWSCognitoIdentityProviderClientBuilder;
 import com.fasterxml.jackson.databind.Module;
 import com.fasterxml.jackson.databind.module.SimpleModule;
-import com.nimbusds.jose.jwk.JWK;
-import com.nimbusds.jose.jwk.JWKSet;
-import com.nimbusds.jose.jwk.KeyUse;
-import com.nimbusds.jose.jwk.RSAKey;
-import com.nimbusds.jose.jwk.source.ImmutableJWKSet;
-import com.nimbusds.jose.proc.SecurityContext;
 import com.revature.rideforce.user.beans.Office;
 import com.revature.rideforce.user.beans.User;
 import com.revature.rideforce.user.beans.UserRole;
-import com.revature.rideforce.user.config.CognitoConfig;
-import com.revature.rideforce.user.config.JWKConfig;
 import com.revature.rideforce.user.json.Active;
 import com.revature.rideforce.user.json.EnumLikeDeserializerModifier;
 import com.revature.rideforce.user.json.EnumLikeSerializerModifier;
@@ -45,8 +28,6 @@ import com.revature.rideforce.user.json.LinkSerializerModifier;
 import com.revature.rideforce.user.repository.OfficeRepository;
 import com.revature.rideforce.user.repository.UserRepository;
 import com.revature.rideforce.user.repository.UserRoleRepository;
-
-import net.minidev.json.JSONObject;
 
 @SpringBootApplication
 @EnableDiscoveryClient
@@ -129,7 +110,6 @@ public class UserApplication implements InitializingBean {
 			admin.setContactInfo(new HashSet<>());
 			admin.setActive(Active.ACTIVE);
 			admin.setRole(userRoleRepository.findByTypeIgnoreCase(ADMIN));
-			admin.setPassword("password"); 							//within setPassword is where we can encode it (more modular)
 			admin.setStartTime((float) 9.0);
 			userRepository.save(admin);
 		}
@@ -139,38 +119,5 @@ public class UserApplication implements InitializingBean {
 	@Scope("prototype")
 	public Logger produceLogger(InjectionPoint injectionPoint) {
 		return LoggerFactory.getLogger(injectionPoint.getMember().getDeclaringClass());
-	}
-	
-	@Bean
-	public AWSCognitoIdentityProvider produceCognitoClient(CognitoConfig cc, JWKConfig jc) {
-		return AWSCognitoIdentityProviderClientBuilder.standard()
-				.withCredentials(new AWSStaticCredentialsProvider(new BasicAWSCredentials(cc.getAccessKey(), cc.getSecretKey())))
-				.withRegion(cc.getRegion())
-				.build();
-	}
-	
-	@Bean
-	public ImmutableJWKSet<SecurityContext> produceJWKSource(CognitoConfig cc, JWKConfig jc) throws IOException, ParseException {
-		// Generate our signing key
-		RSAKey key = new RSAKey.Builder(jc.getModulus(), jc.getPublicExponent())
-				.privateExponent(jc.getPrivateExponent())
-				.keyUse(KeyUse.SIGNATURE)
-				.keyID(jc.getId())
-				.build();
-		
-		// Get Cognito public keys
-		List<JWK> l = new ArrayList<>(JWKSet.load(cc.getJwks()).getKeys());
-		
-		// Add our signing key to the list
-		l.add(key);
-		
-		// Return the complete JWKSet
-		return new ImmutableJWKSet<SecurityContext>(new JWKSet(l));
-	}
-	
-	@Bean
-	public JSONObject producePublicJWKSet(ImmutableJWKSet<SecurityContext> jwkSet) {
-		// Return the public JWKSet
-		return jwkSet.getJWKSet().toPublicJWKSet().toJSONObject();
 	}
 }
