+++ src/main/java/com/revature/rideforce/user/config/CognitoConfig.java
@@ -3,27 +3,20 @@ package com.revature.rideforce.user.config;
 import java.net.URL;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.regions.Regions;
+import com.amazonaws.services.cognitoidp.AWSCognitoIdentityProvider;
+import com.amazonaws.services.cognitoidp.AWSCognitoIdentityProviderClientBuilder;
 
 @Configuration
 @ConfigurationProperties("aws.cognito")
 public class CognitoConfig {
 	private URL jwks;
-	private String poolId;
 	private Regions region;
 	private String clientId;
-	private String accessKey;
-	private String secretKey;
-	
-	public String getPoolId() {
-		return poolId;
-	}
-	
-	public void setPoolId(String poolId) {
-		this.poolId = poolId;
-	}
 	
 	public Regions getRegion() {
 		return region;
@@ -40,22 +33,6 @@ public class CognitoConfig {
 	public void setClientId(String clientId) {
 		this.clientId = clientId;
 	}
-
-	public String getAccessKey() {
-		return accessKey;
-	}
-	
-	public void setAccessKey(String accessKey) {
-		this.accessKey = accessKey;
-	}
-	
-	public String getSecretKey() {
-		return secretKey;
-	}
-	
-	public void setSecretKey(String secretKey) {
-		this.secretKey = secretKey;
-	}
 	
 	public URL getJwks() {
 		return jwks;
@@ -65,9 +42,11 @@ public class CognitoConfig {
 		this.jwks = jwks;
 	}
 	
-	@Override
-	public String toString() {
-		return "CognitoConfig [jwks=" + jwks + ", poolId=" + poolId + ", region=" + region + ", accessKey=" + accessKey
-				+ ", secretKey=" + secretKey + "]";
+	@Bean
+	public AWSCognitoIdentityProvider produceCognitoClient(AWSStaticCredentialsProvider acp) {
+		return AWSCognitoIdentityProviderClientBuilder.standard()
+				.withCredentials(acp)
+				.withRegion(region)
+				.build();
 	}
 }
