+++ src/main/java/com/revature/rideforce/user/config/JWKConfig.java
@@ -1,9 +1,23 @@
 package com.revature.rideforce.user.config;
 
+import java.io.IOException;
+import java.text.ParseException;
+import java.util.ArrayList;
+import java.util.List;
+
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+import com.nimbusds.jose.jwk.JWK;
+import com.nimbusds.jose.jwk.JWKSet;
+import com.nimbusds.jose.jwk.KeyUse;
+import com.nimbusds.jose.jwk.RSAKey;
+import com.nimbusds.jose.jwk.source.ImmutableJWKSet;
+import com.nimbusds.jose.proc.SecurityContext;
 import com.nimbusds.jose.util.Base64URL;
+
+import net.minidev.json.JSONObject;
 
 @Configuration
 @ConfigurationProperties("jwt.registration.rsa-key")
@@ -41,5 +55,30 @@ public class JWKConfig {
 	
 	public void setPrivateExponent(Base64URL privateExponent) {
 		this.privateExponent = privateExponent;
+	}
+	
+	@Bean
+	public ImmutableJWKSet<SecurityContext> produceJWKSource(CognitoConfig cc) throws IOException, ParseException {
+		// Generate our signing key
+		RSAKey key = new RSAKey.Builder(modulus, publicExponent)
+				.privateExponent(privateExponent)
+				.keyUse(KeyUse.SIGNATURE)
+				.keyID(id)
+				.build();
+		
+		// Get Cognito public keys
+		List<JWK> l = new ArrayList<>(JWKSet.load(cc.getJwks()).getKeys());
+		
+		// Add our signing key to the list
+		l.add(key);
+		
+		// Return the complete JWKSet
+		return new ImmutableJWKSet<SecurityContext>(new JWKSet(l));
+	}
+	
+	@Bean
+	public JSONObject producePublicJWKSet(ImmutableJWKSet<SecurityContext> jwkSet) {
+		// Return the public JWKSet
+		return jwkSet.getJWKSet().toPublicJWKSet().toJSONObject();
 	}
 }
